refactor: mypage의 types 파일 추가
