+++ src/apis/record/types.ts
@@ -9,14 +9,3 @@ export interface StudyRecord {
   imageIds: number[];
   writtenAt: string; // '2023-07-27T15:21:09.332Z'
 }
-
-export type UserData = {
-  name: string;
-  time: string;
-};
-
-export type PageType = 'PLACE' | 'RECORD';
-
-export type ViewType = 'BOOKMARKS' | 'REVIEWS' | 'TIMELINE' | 'MAP';
-
-export type CalendarType = 'MONTH' | 'WEEK';
