+++ src/app/mypage/types.ts
@@ -1,0 +1,10 @@
+export type UserData = {
+  name: string;
+  time: string;
+};
+
+export type PageType = 'PLACE' | 'RECORD';
+
+export type ViewType = 'BOOKMARKS' | 'REVIEWS' | 'TIMELINE' | 'MAP';
+
+export type CalendarType = 'MONTH' | 'WEEK';
